+++ packages/fxa-auth-server/lib/routes/subscriptions/mozilla.ts
@@ -74,29 +74,36 @@ export class MozillaSubscriptionHandler {
       email,
       'mozillaSubscriptionsCustomerBillingAndSubscriptions'
     );
+    try {
+      const stripeBillingDetailsAndSubscriptions =
+        await this.stripeHelper.getBillingDetailsAndSubscriptions(uid);
+      const iapGooglePlaySubscriptions =
+        await this.playSubscriptions.getSubscriptions(uid);
 
-    const stripeBillingDetailsAndSubscriptions =
-      await this.stripeHelper.getBillingDetailsAndSubscriptions(uid);
-    const iapGooglePlaySubscriptions =
-      await this.playSubscriptions.getSubscriptions(uid);
+      if (
+        !stripeBillingDetailsAndSubscriptions &&
+        iapGooglePlaySubscriptions.length === 0
+      ) {
+        throw error.unknownCustomer(uid);
+      }
 
-    if (
-      !stripeBillingDetailsAndSubscriptions &&
-      iapGooglePlaySubscriptions.length === 0
-    ) {
-      throw error.unknownCustomer(uid);
+      const response: {
+        customerId?: string;
+        subscriptions: MozillaSubscription[];
+      } & Partial<PaymentBillingDetails> = stripeBillingDetailsAndSubscriptions || {
+        subscriptions: [],
+      };
+
+      return {
+        ...response,
+        subscriptions: [
+          ...response.subscriptions,
+          ...iapGooglePlaySubscriptions,
+        ],
+      };
+    } catch (err) {
+      console.log(err);
+      throw err;
     }
-
-    const response: {
-      customerId?: string;
-      subscriptions: MozillaSubscription[];
-    } & Partial<PaymentBillingDetails> = stripeBillingDetailsAndSubscriptions || {
-      subscriptions: [],
-    };
-
-    return {
-      ...response,
-      subscriptions: [...response.subscriptions, ...iapGooglePlaySubscriptions],
-    };
   }
 }
